requirements.txt
tensorflow-cpu==2.13.1
keras==2.13.1
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
numpy==1.24.3